InputManagerをユーザと共有できるフォルダに移動
