+++ Engine/Core/src/yougine/Editor/InspectorWindow.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 #include "EditorWindow.h"
-#include "../InputManager.h"
+#include "UserShare/InputManager.h"
 #include "ComponentViewer.h"
 #include "SelectionInfo.h"
 #include "../component_factory/ComponentFactory.h"
